+++ packages/yourloops/pages/notifications/notifications-list.tsx
@@ -21,7 +21,7 @@ import { AppBar, Breadcrumbs, Container, createStyles, Link, List, ListItem, mak
 import HomeIcon from "@material-ui/icons/Home";
 import NavigateNextIcon from "@material-ui/icons/NavigateNext";
 
-import HeaderBar from "../../components/header-bar";
+import HeaderBar from "../../components/primary-header-bar";
 import { INotification, Notification, NotificationType } from "./notification";
 import { UserRoles } from "../../models/shoreline";
 import { useAuth } from "../../lib/auth";
