+++ packages/yourloops/test/pages/hcp/index.ts
@@ -26,7 +26,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import testPatientsListAppBar from "./patients-list-bar.test";
+import testPatientsSecondaryBar from "./patients-secondary-bar.test";
 import testPatientListTable from "./patients-list-table.test";
 import testTeamCard from "./team-card.test";
 import testTeamEditDialog from "./team-edit-dialog.test";
@@ -34,12 +34,12 @@ import testTeamAddMemberDialog from "./team-member-add-dialog.test";
 import testTeamRemoveMemberDialog from "./team-member-remove-dialog.test";
 import testTeamSwitchRoleDialog from "./team-member-switch-role-dialog.test";
 import testTeamMembersTable from "./team-members-table.test";
-import testTeamNavBar from "./teams-nav-bar.test";
+import testTeamsSecondaryBar from "./teams-secondary-bar.test";
 import testTeamPage from "./teams-page.test";
 
 function testHCPPage(): void {
   describe("Patient List", () => {
-    describe("AppBar", testPatientsListAppBar);
+    describe("AppBar", testPatientsSecondaryBar);
     describe("Table", testPatientListTable);
   });
   describe("Teams", () => {
@@ -49,7 +49,7 @@ function testHCPPage(): void {
     describe("Remove member dialog", testTeamRemoveMemberDialog);
     describe("Switch role dialog", testTeamSwitchRoleDialog);
     describe("Members Table", testTeamMembersTable);
-    describe("Nav bar", testTeamNavBar);
+    describe("Nav bar", testTeamsSecondaryBar);
     describe("Page", testTeamPage);
   });
 }
